Define std::hash for HexCoord.
Add a template specialisation of std::hash for HexCoord, so that it can be
used efficiently as the key for std::unordered_map and the likes.

Since we just have two small integers as coordinates, combining them into
a value of size_t is typically even possible without any collisions at all.