Add dummy BaseMap class for now.
Define a BaseMap class, which will be used to hold the underlying
map data like obstacles and the tiles-to-region map.  For now, it
just returns a dummy edge-weight function until we get the actual
map data integrated.

This will allow us to implement the core logic for the findpath RPC
and other stuff (e.g. characters moving around) right now, without the
need to have the actual map data integrated already.