Introduce libpx.la for easier testing.
When building pxd, split off some files (basemap for now) into a private
library libpx.la.  This will allow linking to those files for easier
unit tests in the future.  The only files that will remain directly in
pxd are those that are not tested and just glueing stuff together
(like main.cpp and perhaps the pxrpcserver).