Pass --static for pkg-config.
Some dependency libraries (e.g. gflags) may be available only statically,
so that we have to link them statically instead of dynamically.  In that
case, we need to pass --static to their pkg-config invocation to ensure
that all of their dependencies are listed correctly.

This is a bit hacky, but seems to be the best solution.  See also
https://bugs.freedesktop.org/show_bug.cgi?id=19541.