Add character-state protocol buffers.
Include a new folder "proto" for protocol buffers.  For now, that holds
two simple files that define hex coordinates and a first version of the
character state (only to support moving around in the future as a first
step).  Also includes integration of the protocol buffer compilation with
the autotools build system.