Test repathing from intermediate points.
Add another unit test for the path finder that verifies that the exact
same path is found if we re-calculate the shortest path from intermediate
points along a previously calculated, longer path.

This ensures that we get consistent movement for characters between their
current position and the next waypoint, independently of how and when
path-finding happens.