Create Character instances through the table.
Make the constructors of Character private, so that instances have to
be created through the CharacterTable.  This moves all client code from
direct Character instances over to a "handle", which is just a unique_ptr.
This makes the instances that are used moveable and destructable at will
without the need for arbitrary code blocks.

Also renames GetAll and GetForOwner to Query*, since those do not
actually return a Handle but instead a Database::Result.  Get* methods
will be used only for thigns that return a Handle:  Currently, only
GetFromResult.  In the future, also GetById.