Implement real game-state to JSON function.
Add a real implementation for the game-state to JSON function, so
that getcurrentstate() now actually returns a list of created characters.

This does not include all data, e.g. the proto movement field is not
covered.  But that will be added later as actually implemented.

Also, there are not yet any custom RPCs to retrieve just characters owned
by a certain name.  Those will also be added later as needed.