Avoid updating proto data if not necessary.
Track the "dirtyness" of the proto data in a Character separate from
that of other fields.  This allows us to avoid updating the proto binary
BLOB unless it was really changed.

The benefit of that is that many frequent updates (e.g. most of the time
while moving) can be done in the fields only, and do not require an
update of the big binary proto blob.  Avoiding that update makes sure
that the undo data can be smaller.