Update data schema for movement design.
Change the Character proto and database structure (as well as the wrapper
class) for the final design of movement.  In particular, we have a
precomputed path to the first waypoint in addition to the list of
waypoints.  This helps us make movement work if obstacles can be dynamic
(e.g. buildings).

Also add a database field for "partial steps", where we accumulate movement
for speeds that are slower than one tile per block.