Test stepping of full L1 range with PathFinder.
Add a test that computes a path that uses the full L1 range (but is not
out of range) and also steps it through.  This was a previously uncovered
special case.

Also adds a quick return when source and target are further away than the
L1 range in the first place (and modifies the corresponding test to verify
this behaviour).  Since this makes the PathHitsTarget benchmark
pointless, it is removed as well.