Use flat vector for RangeMap.
Replace the previous std::unordered_map in RangeMap with a flat vector
that stores the L1 range around the centre directly.

This allows us to take advantage of the specific map structure, and avoid
costly hashmap inserts / lookups (which were the bottle neck for path
finding previously according to callgrind).  This improves the benchmark
for PathToTarget/100 from 27ms to 13ms.