Do not include immutable values in UPDATE.
Exclude immutable properties of Characters (name and faction) from the
database UPDATE statement, as it is not necessary to overwrite them
there.  They are still included in the INSERT OR REPLACE that is used
for the initial creation (where it is necessary) and also for the proto
update (where it is just an "accident" from the structure).