Use a temporary table for L1 range filtering.
Instead of simply filtering on `x` and `y`, use a temporary table and
JOIN when looking for database rows that have coordinates in some given
L1 range.  This allows full usage of the INDEX on (`x`, `y`), rather than
just using it for the `x` filtering and doing a full scan on `y`.

DO NOT SUBMIT - no unit tests yet, and it still seems to be slower