Database abstraction for fighters.
Fighters are entities that can attack others and can take damage.  In other
words, those will be characters and buildings.

This creates an abstraction layer that makes both types look the same for
the combat code.  For now, we only have characters anyway, but buildings
can then be added easily in the future.