Character utilities for Python gametests.
Add some (better organised) utilities for working with characters from
the Python gametests.  They can now be created easily with a new function,
and a new class wraps existing characters so they can be sent and
inspected easily.

This obsoletes some custom code in the existing tests, so that they are
simplified a bit.