Specify attacks through the character proto.
This introduces a new CombatData proto, which is part of the character
state proto and holds, for now, a list of attacks.  Instead of hardcoding
a range of 10, target selection now looks at the actual attacks that
are there and figures out the range it needs to check from there.

For now, each created character still just gets hardcoded attacks set;
but the combat code itself does not need to know about that anymore.