Add HP to characters.
Define a new HP message in the combat protos.  The CombatData of a
character defines the maximum HP values as well as HP regeneration.
The database table for characters also gained a new column to keep the
"frequently updated" current HP in a separate proto BLOB column.

The Character handle class also supports retrieving / modifying
the HP proto of a character (and updating it back to the DB).

We also initialise the HP and stats when a character is created
through the MoveProcessor.  For now, they are just set to some
"random" values.