Track killed characters when dealing damage.
Keep track of what fighters have been killed while dealing damage and
return all their target IDs from DealCombatDamage.  This can be used
in a later step to actually process the kills (which is something
specific to characters and in the future buildings, so cannot be done
on the fighter abstraction level).