Use explicit 64-bit-int functions for JSON.
This replaces all "isInt"/"asInt" and UInt functions in jsoncpp with
the explicit 64-bit variants.  That ensures that we only extract values
which will fit into 64-bit integers, independent of the system (as long
as it has 64 bits at least).

Doing that prevents potential forking issues with integers that may
overflow int/unsigned in some systems but not others.