Remove compact writing of obstacle data.
Cleans up the code to write obstacle data in "compact" binary form.  Since
we always use the code output to embed the data into tauriond, there is no
need for the unused compact writing code for now.

We will add output of the region map as compact binary in the future (since
it is too large to be embedded into the binary), but that will need its
own code path and code anyway.