Output general tiledata generated C++ code.
For the generated C++ code that contains static map data to be built into
the binary, output a general "tiledata" file rather than "obstacles" and
restructure the namespaces so that min/max X/Y is "independent" of the
obstacle data.

This is in preparation for adding more stuff, namely metadata for the
region map, to this in the future.