Add raw input data for region map.
This adds the raw input data for the tile-to-region map, similar to
the input data for obstacles.

The data itself is not checked in (just like the obstacle data is not),
but expected to be placed in the "data" folder.  The makefile will then
uncompress it, and the resulting output files are .gitignored.

The data is not yet processed in any other way (except being
uncompressed).