Do not dist the data folder.
Remove files in the data folder from the dist target (and the data folder
itself from automake completely).  Those files will be dist'ed from
mapdata, where symlinks are placed.

The data folder is just for convenience in the repository, so that all
those files not checked in are in a single place.