Define starting positions for each faction.
Instead of starting every character in (0, 0), we define a distinct
starting location for each faction.  The positions have been chosen
by Konstantin.

This is still for testing only for now and will be changed later.  But it
already ensures that characters are not directly attacking each other
after creation.

Includes additions to the integration tests so that they are independent
of the starting positions (and thus still work after this change).