Implement GetRegionShape.
The new method RegionMap::GetRegionShape can be used to get not just the
region ID of some coordinate, but also the shape of the region around it
(i.e. all tiles of the same region).

This is something not directly needed by the game backend (presumably),
but exposed through an RPC to the frontend.