Use x ranges for faster GetRegionShape.
Instead of processing tiles one by one for the flood fill in GetRegionShape,
we can make use of the compact data format encoding the region map to get
ranges of x coordinates of tiles in the region "for free".  With this, we
can optimise GetRegionShape.

The speedup is limited, though (64us -> 57us) and it hurts GetRegionId
a little (310ms per 1M lookups -> 340ms).  Thus it seems not worthwhile
to apply this in general.