Remove character names.
This removes the concept of a "character name" entirely from the
database schema and code.  Character names are just a burden and do
not really yield much game benefit.  Characters will be associated to
owner names, those will be the "names" shown on the map.

This fixes https://github.com/xaya/projectx_backend/issues/3.