Add busy state and prospection to characters.
This adds a new field to the character table, which tracks how many
blocks each character is still "busy" (e.g. prospecting).  Details
about the ongoing operation are stored in the character's proto, which
for now only supports prospecting.

The CharacterTable is expanded as well, so that it can automatically
decrement the busy counters and give out those characters that have
finished their operation for specific processing.