Memory-efficient map for all tiles.
The new DynTiles container is a map from all tiles on the basemap to
values of arbitrary types (including packed bool's).  Unlike
FullRangeMap, this container uses knowledge about the map's structure,
so that the data is stored efficiently (exactly as many entries in the
underlying std::vector<T> as there are tileson the map).