Class for dynamic obstacle map.
The new DynObstacles class handles the dynamic obstacles on the map
(vehicles in the beginning, buildings later).  Instances are initialised
from the database on construction (once per block update).  They will be
updated as characters are created and vehicles move around during the
block update in the future (as well as actually being used as obstacles).