Move character speed to proto field.
Instead of hardcoding the movement speed of characters in the movement
code itself to 750, move that to a field in the character proto and use
the value from there.

The speed itself is not changed and still at 750.  But now that value is
just set from Params when initialising the stats of a new character, which
a) puts the magic value in Params (where it belongs) and b) makes it easy
to modify the speed later when we have e.g. different types of vehicles.