Handle random spawn locations well in tests.
This upates the integration tests to make them more robust with respect
to the positions of characters.  With this change, random spawning
locations of characters are handled well.

This also includes a change to the actual spawn locations, so that they
are in a region more suitable (with fewer obstacles and no isolated tiles
around them).