Make vehicles obstacles for movement.
Consider vehicles (of the same faction) as obstacles during the movement
phase, to prevent stacking also there.  This also includes proper updating
of the dynamic obstacle map for vehicles that are moved, so that it is
correct when processing each individual vehicle.

This implements https://github.com/xaya/taurion_gsp/issues/4.