Wrap partial step into proto.
Instead of tracking the "partial step" value directly as a database
column, wrap it into a new "VolatileMovement" proto (similar to the
current HP one).

This allows to still update the value without touching the big general
proto, but it will enable us to add more volatile movement fields in the
future.  In particular, we will track the number of failed stepping trials
for characters that are blocked by a dynamic obstacle.