Retry blocked steps a couple of times.
When a character tries to step into an obstacle (along already planned
steps), it did stop completely so far.  With this change, we give it
a couple of turns (10 at the moment) where it retries that step.  If the
way is free on any of the retries, then movement continues.

This helps with situations where only another character passing by in
the front was blocking the path.  But due to the limit on the number
of retries, we also make sure that we do not end up with many characters
being blocked forever but still leading to movement computations on
each turn.