Introduce framework for computing fame.
This adds the FameUpdater class, which will be used to compute/update
fame for kills in the game (based on the damage lists).  For now, this
is mostly a refactoring and a utility for some more tests (in particular
the handling of characters that killed another one and are killed
themselves in the same turn).  There is not yet an actual computation
of fame updates, as the rules for that are not yet fixed.  But this will
be easy to fill in later.