Allow creating multiple characters at once.
This modifies the "create character" move so that the "nc" value
should be an array of objects instead of a single object.  If there
are multiple entries, then all of them are processed, so that a single
transaction / name_update can create multiple characters (if enough
developer payment is attached).