Database table and wrapper for account data.
This creates a new database table, which will hold data associated to
accounts (Xaya names).  For now, the data is fame and number of kills,
but in the future it could hold also vCHI balances and other stuff we
need on a per-account basis (rather than per-character).

Also includes a database wrapper class for these accounts, to be used
for access from the game-logic code itself.