Compute fame with deltas.
Rewrite the fame computation, so that it does not keep original fame values
and updates on the fly, but instead it computes deltas that are only
applied at the end.

This solves, in particular, an issue where fame could underflow zero when
a name with fame 100 had two characters killed at once.  It also removes
potential order-dependence with applying the caps at 0 and 9999, since
the computation in-between is now done without any caps and those are applied
only at the very end.