Factor out some move-parsing code.
This introduces BaseMoveProcessor as a new base class for MoveProcessor.
Some of the logic for parsing and validating moves is moved there, so that
we can reuse it for processing of pending moves in the future.

For now, this is just a refactor without changes to the behaviour in any
way (and mostly consists of moving around code with small tweaks).