Use "buckets" to make DynTiles sparse.
Instead of allocating a full array of values with an entry for each map
tile, allocate instead an array of "buckets" where each bucket corresponds
to a fixed number of tiles (2^16 in this commit).  This allows us to allocate
a bucket only when needed, which improves memory usage and especially
constructor performance for DynTiles while the map is empty and we have
mostly default values anyway.

Compared to the original implementation, this cuts down constructor
(and destructor) time massively from 2086us to 1.38us, while other operations
(accessing and updating entries) is slowed down only moderately, by 30-50%.