Stop giving out prizes after competition end.
This introduces the official end time of the tech demo competition
(at UNIX timestamp 1571148000) and stops giving out prizes in blocks
that are beyond that timestamp.

Note that this causes a hard fork of the game logic:  Because we stop
giving out prizes, we also try fewer random numbers, and thus alter
the entire sequence of random numbers.  In any case, the fork only
applies at the competition end, so it won't matter for the competition
itself.

This implements https://github.com/xaya/taurion_gsp/issues/25.