Order character updates in a move by ID.
Explicitly order the updates of characters within a single move by the
character ID as integer.  Previously, we relied on JsonCpp's iteration
order, which was also sorted but by string.  In this change, we make the
order explicit in our own code (rather than depending on JsonCpp) and
also change it to the more logical order by integer ID.

Note that this is a consensus fork, although unlikely to actually cause
differences in practice.

This fixes https://github.com/xaya/taurion_gsp/issues/27.