Explicitly initialise accounts with moves.
This adds a new move for initialising an account (and setting its faction)
to the core game.  All the tests are updated to initialise accounts
before creating characters as needed as well.

For now, the account faction is not tied to the character faction, though.
Removing the explicit faction for created characters and using the faction
from the account will be the next step.