Ignore moves from uninitialised accounts.
Early stop move processing from accounts that have not been
initialised (after checking for the initialisation itself).

This ensures that we won't get into a weird situation where perhaps
a move is actually valid to process but the account is not initialised
yet.  It also forces us to fix some more tests that were relying on
not-yet-initialised accounts (mostly in move processing).