Add getnullstate RPC method.
The new getnullstate RPC method just returns JSON null as payload data,
but in addition also the current syncing state and other default stuff
from libxayagame.  This makes it useful for checking if the GSP is
synced up, without the need to retrieve any other (potentially expensive)
data from it.