Add typing to Database::Result.
Instead of setting a string name on Database::Result to check that the
result is actually from the table / query we expect, add a type as template
parameter to it.  The type specifies what kind of result it is (e.g. a row
from the characters table or a row from regions).

This makes the calling code more robust, as we now get static type checking
(as well as explicit documentation of the expected types) for results.

As a follow-up step, we can then also use the result types to hold more
information, like the expected columns.  That will allow us to verify those
also more thoroughly, and help with performance by changing column
extraction by string name into extraction by an integer column (which can
be looked up more quickly).