Define columns in ResultType.
Add explicit definitions of the expected columns to Database::ResultType
classes.  This helps us remove redundancy, as the type is already known
when we call Get or GetProto.  And it also replaces the column lookup
by string name with a lookup by index into std::array, which should
make the code faster overall.