Parse database protos lazily.
This introduces the LazyProto template, which is a wrapper around a
protocol buffer message.  Database operations work on this instead of
the raw protos.

The LazyProto class allows us to parse protocol buffers lazily, only
when they are actually accessed in game logic.

For now, this has no impact on real-world syncing performance, because
most code accesses the general character proto in one way or another.
But with this change in, we have more control over proto parsing in general,
so it is useful to have nevertheless.  And going forward, we can focus
on adding some important fields to other places, and avoid accessing
the general proto in many cases.