Split out regeneration data proto.
Instead of having the HP regeneration data (max HP and regeneration rate)
as part of the general character proto, split it out into its own proto
and database column.

While this data is not changed often, it is accessed for every character
on the map during HP regeneration all the time.  Thus it helps to have it
separate, so that the main proto does not need to be deserialised for
each character for every block.