Gracefully handle empty waypoints in move.
The movement code assumes that the "movement" proto is present only
if there are actually waypoints.  This was not the case when a move
explicitly set the waypoints to [].

With this change, we gracefully support this situation.  A move setting
waypoints to [] is now explicitly supported, and can be used to just stop
a character's movement.