Use custom query for filling DynObstacles.
Add a custom function to query just for positions and factions of
characters on the map to the CharacterTable.  Using this, we can
more efficiently fill in DynObstacles, without the need to construct
full-blown Character wrappers for everyone on the map in the process.

This makes sync of the tech-demo competition blocks faster by about 4%,
and reduces the typical time per block for them by 1-2ms.