Avoid parsing character proto for combat targets.
This adds a character's attack range to the database (computed whenever
the main proto is modified).  With this and the already existing
"has target" field, we can perform target finding without accessing the
main character proto at all.

This speeds up target finding, and reduces the total amount of time
spent in protocol buffer parsing to below 1% of the entire runtime.
It improves syncing speed for the tech-demo competition blocks by
about 1-2ms per block.