Disallow out-of-range Get for RangeMap.
Instead of supporting out-of-range read-only Get in the RangeMap
(which returned the default value), explicitly forbid this and update
the code to not rely on that feature.

In this way, we can later on remove the range check entirely for
"release" builds if we want.  This should give a good speed boost.