Add option to enable slow assertions.
This adds a new option "--enable-slow-asserts" to the configure script.
If enabled, we turn on assertions in the code that slow down syncing
noticably (and should thus be off for production builds).

For now, we use this to add back more tests in character validation,
that counteract the laziness introduced for performance reasons.  Later,
we will also make some of the checks in the most critical places of
BaseMap and DynTiles dependant on slow assertions to actually speed up
path finding further.