Only check "to" tiles for obstacles.
When checking edges during path finding, only consider them impassable
if the "to" file is an obstacle.  This avoids half the calls for basemap
and dyntiles "IsPassable".  It does not matter in practice, except for
situations where someone would already be on top of an obstacle (which
should not happen anyway unless there is another bug).

Note that the stepping code previously had a bug, which assumed symmetric
edge weights / obstacles.  Now we check the individual edges again during
stepping, which means that it works correctly as it should even for
unsymmetric edge-weight distances.