Allow updating characters with an ID list.
Extend move processing for character updates ("c" move field) to support
ID lists instead of single IDs.  If an ID list is given, we update all
characters from that list to the same value.  This can be useful for
moving groups of characters, for instance.

This also removes the test for duplicated JSON keys.  libxayagame now
has an explicit test for that, and is supposed to do the dedup (by means
of JsonCpp) already before passing on the data to the GSP.  Thus there
is no need to worry about this at all in Taurion.

Implements https://github.com/xaya/taurion_gsp/issues/31.