Add chosen speed to movement proto.
The new field "chosen_speed" in a character's movement proto can
be used to set a desired speed.  If set, the character will travel
with the slower of the chosen speed and its own intrinsic speed.
This is useful for moving as a convoy.

For now, the field is only implemented in the movement code and tested
in unit tests.  Later on, it will be able to set it through moves.