Remove faction from character creation moves.
When creating a character, expect the command to just be an empty object
{}, without a faction.  The faction is instead taken from the account
corresponding to the move.

So character creation moves now look like this:

  {"nc": [{}, {}, {}]}

In the future, we might want to include some other properties for the
new characters in the {} objects.  For now we don't need any extra data.