Create ResultWithCoord.
The new database/coord source file defines a basic Database::ResultType
for results that contain coordinates (x and y columns).  For now, this
is just the character, but it will also be e.g. loot on the ground in the
future.  By moving out the logic for handling coordinates in the database,
we can support those things easily in the future.