Add ground loot to database.
This creates a proto definition for a general "inventory" (of a character,
an account in a building or loot on the ground).  It also defines a simple
wrapper class that takes care of managing it in an intelligent way (e.g.
determining when items are gone and can be removed entirely, rather than
setting the count explicitly to zero).

Also, this defines a database table and access handles for loot on the
ground; this is essentially an inventory keyed by hex coordinates.