Add ground loot to game-state JSON.
Return information about ground-loot piles in the game-state JSON.  The
data is part of a new field "groundloot" in the full game state, and it
is also returned by a new specific split RPC "getgroundloot".

This also defines a general inventory-to-JSON routine which will be
shared for character inventories in the future.