Use uint64_t for chosen speed.
When parsing the chosen speed from a move, store the intermediate value
as uint64_t rather than unsigned.  We do a range check on it anyway (and
then just allow values up to one million), but we need to make sure the
value gotten from JSON can be stored in the mean time.