Introduce Inventory::QuantityT.
Add a specific type to hold the quantity of items (aliased to int64_t).
Also add a constant for the maximum valid amount and check values in the
god-mode drop command against it.

This will be more useful for character commands, as it is an early check
to avoid overflows triggered by malicious users.