Set character cost to 1 CHI.
For testing between now and the start of the second competition, put the
CHI cost of characters down to 1 CHI.  This makes it easier to test on
mainnet when we work on finalising the next version (0.2).

Before the competition actually starts, we will reset the cost back up
to a higher number.