Add read-only config data proto.
This defines a new protocol buffer, which holds "read-only, static config
data" for Taurion.  For now, it contains data of known item types and
their cargo-space usage.  In the future, we can also add more data for
item types as well as e.g. define vehicle stats and other things.