Add cargo space limit for characters.
This defines a cargo-space limit for characters (set to 1,000 for now
as an arbitrary number).  It is verified in the slow assertion checks
when destructing a Character instance, and all tests are adapted to set
it accordingly in order to still work.

The limit does not yet determine what items can be picked up, though, so
it is currently possible to trigger that CHECK-fail deliberately by
just picking up more stuff than the available cargo space.