Implement area attacks.
Attacks can now also be "area-based" rather than the previous type of
target-based attacks.  In that case, all enemies in range take damage,
not just the randomly-selected target.

We still do target selection in the same way (even if an attacker has only
area attacks), as the presence or absence of a target tells us whether or
not there are any enemies in range.  But the actually selected target
does not matter if the attacker has only area attacks.

The default attacks are changed such that the short-range one is area
based now.