Remove extra prizes.
Remove the extra prizes for the first demo competition (we will set
fresh prizes accordingly before starting the next competition).  Also
make bronze have one available prize (which is guaranteed to be won on
the first trial that doesn't win any other prize) instead of none.  With
this, we can still easily test prize exhaustion, but we can also test
prospecting more easily with prizes as items since the first attempt is
guaranteed to yield an item.