Add prospected height to region data.
When a region is prospected, store also the height at which prospecting
finished with it.  This will be used later to allow re-prospecting after
all resources have been mined and a certain time interval has passed.

We might also want to add a database index on this later and add an RPC
that queries for all regions prospected only recently.