Add mining data to character proto.
Add a member to the character proto that will hold mining data
(supported rate of mining and whether or not mining is active).

The default characters will mine with rate one unit per block for now.

The new mining data is also returned in the game-state JSON for characters
in a new "mining" JSON field.