Implement processing of mining.
Add a step during block processing that actually performs mining with
characters that have started it.  This step moves mine-able resources
from the region to the character inventory, as long as there is still
something left and the cargo is not full.  If nothing can be mined in
one turn, then we stop mining.