Implement proper resource distribution.
This replaces the dummy distribution of resources that we used so far
with a "real" one that we will use in the next competition.

We have a list of areas where certain types of resources can be found,
and then randomly choose resources based on the current position, how
far it is to those areas, and what the "base amount" for each resource
type is.

See https://github.com/xaya/taurion_gsp/issues/53 for more details.