Pass character ID through ProcessAllPositions.
CharacterTable::ProcessAllPositions can be used to efficiently
process all character positions on the map, without the need to
retrieve and/or parse the full character data (mostly big protos).

This is used for updating the dynamic obstacle map, but it can also
be used for general "selective processing" of characters based on
their position, e.g. for banking.  In order to allow that, however, we
need to pass through the character ID as well, so that the characters
that should be processed can.