Implement banking.
This implements banking as described in
https://github.com/xaya/taurion_gsp/issues/60.

When a character is inside a designated banking area at the end of block
processing, then all its inventory is moved to the account's "banked"
inventory.  And if there are enough resources to complete one or more
predefined "sets", then those are removed and instead banking points added
to the account.