Adjust character stats per faction.
Set adjusted character stats, which are differentiated between the
factions.  These will be the ones we use for the second competition
(unless testing shows we need more tweaks), and have been designed by
Andy.

Note that they make some of the integration tests fail, which will be
adjusted in a follow-up.