Remove competition end time for prizes.
For the first competition, we introduced an "end time" in the GSP, after
which no prizes could be found anymore.  For the second competition, this
is mostly irrelevant, since banking the prizes is what counts (not
finding them).

For simplicity this change removes the end-time concept completely.  We
will use a snapshot of the exact game state at the competition end block
anyway to determine the outcome.