Disallow mining when a character is busy.
Don't allow to start mining with a character that is busy.  This has not
really an effect for now, as the only busy is prospecting and you can't
start mining anyway until the area is prospected, but it is probably good
to disallow this anyway in case we add more busy later.