Track modified height for regions.
When modifying a region, store the block height into the database.
This allows us to query regions only modified since a given block height,
which we will expose next in the RPC interface.

The amount of regions in the database will grow quite big over time, so
that we should not rely on being able to read and process all regions for
every block in the frontend.