Add bootstrap data RPC.
The new RPC getbootstrapdata returns all regions, and thus the data that
is necessary for the frontend after startup to initialise itself before
it can start with filtered, recently-modified region updates.

In the future, this bootstrap data may contain more fields, and it may
also be served e.g. using a dedicated HTTP interface instead (or in
addition).