Switch to new SQLiteDatabase.
Update the code for the libxayagame refactoring of
http://github.com/xaya/libxayagame/pull/96.

We still keep our custom Database wrapper class, as it also gives us
access to other SQLiteGame functionality (e.g. the AutoId's).  But we
can simplify the test Database implementation, making use of
SQLiteDatabase e.g. for SQLite initialisation or the statement cache.