Encode roconfig as binary proto.
Convert the roconfig text proto to a binary blob during build, and then
embed the binary data into the executable.  This is more efficient.

It also has the benefits of catching errors in the text proto during build
(rather than in unit tests), and it allows passing the binary proto e.g.
to the frontend to share data with it.