Allow characters to be inside a building.
Add the database schema and accessor parts that allow characters to be
inside buildings.  That will also be reflected in the game-state JSON,
and taken into account for various functions that process characters
on the map (e.g. target finding).

There is not yet a way to actually get inside a building (except by
unit tests), and also certain moves would fail to process when issues
by a character inside a building.