Move to specify intent to enter building.
A new move can be used to specify the "intent" to enter a given building
with some character (or to cancel that intent later).  This can be done
any time (except when inside a building), as long as the building ID is
valid and of an enterable building (not opposing faction).

The flag will then stick around until the character can actually enter,
e.g. is no longer busy and comes within reach of the building.

The move has this form (for setting the intent with a given building ID
and for clearing the intent):

  {"eb": 42}
  {"eb": null}

We also return that intent in the character-state JSON, and track the
"enter building" moves while pending.