Allow partial JSON testing for null's.
In the partial JSON comparison, we use null values in the expected JSON
to indicate fields that *should not* be present.  This makes it impossible
to test for fields that *should* be present but have a null value.

This change adds in a special rule, which makes a literal "null" string
compare equal to JSON null, allowing for such tests.