Remove banking.
With the 2nd competition over, we no longer need any of the banking
logic.  That is something that won't be in the full game.  And with
building inventories, we now even have game logic (from the full game)
that could be used to mimic the banking in the competition, in case we
need it for future competitions.