Make target separate database column.
Instead of having the selected target be part of the main character
proto and having a separate indexed column for "has target", just make
the target a column in the database itself (like HP and RegenData proots).
We have an index on this, so we can select by "IS NOT NULL".

This will also allow us to easily reuse the same structure and code
for buildings, as we then don't need to template on the proto type
or something like that.