Use CombatEntity in Fighter.
This extends the CombatEntity abstract class to expose all data
necessary for combat logic.  With this, we can change Fighter to
only use std::unique_ptr<CombatEntity>, instead of the more specific
Character and Building handles.

This is a first step towards getting rid of Fighter, and just using
CombatEntity directly in the code that now uses Fighter.