Store account fame/kills in proto.
Define a proto and BLOB database column for accounts, and store kills
and fame in there rather than dedicated database columns.

These fields are likely not performance critical, so having them in the
proto simplifies the database schema and logic.  Also having a proto in
the first place will allow us to use it for more things in the future,
like vCHI balances or skills.