Move processing for coin operations.
This defines a new move, which can be used to burn one's coins (vCHI)
and/or transfer them to someone else.  The format is like this:

  {"vc": {"b": 123, "t": {"foo": 10, "bar": 20}}}

This example move would burn 123 coins, send 10 to foo and 20 to bar.
The recipients must be initialised accounts.

Burns are processed before transfers, and the transfers are processed in
alphabetical order.  Any invalid operation (or one that would exceed the
balance remaining after the previous oneos) is ignored, and the following
are still tried to be done.