Move processing for service operations.
Integrate service operations with the move processor, so that they
are actually parsed and executed.  For now, only refining of materials
is defined, which can be requested with a command like this:

  {"s": [{"b": 123, "t": "ref", "i": "foo", "n": 3}]}

This command requests to refine 3 units of "foo" inside building 123.
Note that the account need to have sufficient vCHI to cover the fees,
and also the requested refining amount must be a whole multiple of the
"input units" for a "refinement step" of the given item type.